fix: Fix import/export bug in login function
